docs(examples): load Compose from a sibling YAML file
Show run_from_file against a three-service Redis, API, and Ubuntu
Compose file instead of inlined YAML.
